ir::core

Struct IRGenerator

pub struct IRGenerator { /* private fields */ }

A struct for generating LLVM Intermediate Representation (IR) from a module of abstract syntax trees (AST) and symbol table stacks (STS).

Implementations

impl IRGenerator

pub fn new() -> Self

Creates a new IRGenerator instance.

This initializer sets up an IRGenerator instance, setting up its own context, module, and builder which are necessary components of writing IR. The most important to understand is the builder, which performs the function of writing IR into the module you are creating.

The current function is not initialized yet, and is set to None. This will change when you generate a function. Once you do so, you must set this so that you can know where to write IR into for each function. Remember, main() is a function!

sts is not set yet, and is None. Additionally, the sts stack pointer is set to 0. Set the STS when you have one and make sure to increment and decrement the stack pointer as you generate IR to stay in step with the STS.

current_target_stack is initialized as an empty stack. In C, in a while loop for example, break; will end the loop entirely upon reaching that statement, and continue; will jump back to checking the loop conditions. The targets for break and continue are stored here, so you know where to point to when you hit one of these.

current_label_id starts at 0 and is the suffix for labels in IR. Jumping to labels is how we do complex logic like loops, and all labels have to be unique (otherwise how would we know which label was which?) so this along with its supporting functions allow unique IDs to be created for every label.

store is initialized to a new store. The store is a space for managing variables, allowing you to assign and reassign variables in IR while keeping track of their tags.

current_insert_block is set to None, this is a space to hold the “basic block” (labeled area of code) that new blocks should be inserted after. ResourcePools contains a function that can create basic blocks after a given basic block, and this helps support that. This will be necessary for inserting code blocks in the body of a while loop for example.

Returns

Returns a new IRGenerator instance ready to generate a module from an AST and STS generated in previous steps.

pub fn get_resource_pools(&mut self) -> Arc<Mutex<IRManager>>

Returns a protected reference to the resource pools.

pub fn get_context(&self) -> ContextTag

Gets the current IR generation’s context, so that a single module of IR generation occurs within the same scope.

Returns

Returns the current context’s tag for use in other IR generation functions that make use of it.

pub fn get_module(&self) -> ModuleTag

Returns the current module being generated.

Returns

Returns the current module’s tag for use in getting the completed module after IR has been generated.

pub fn get_builder(&self) -> BuilderTag

Returns the builder used for constructing IR statements.

Returns

Returns this IRGenerator’s builder’s tag for use with functions that require the builder.

pub fn get_function(&self) -> Option<ValueTag>

Retrieves the current function being built.

Returns

Returns the tag of the function currently being built for use with functions that need this information, could be None if called before initialized.

pub fn set_function(&mut self, function: ValueTag)

Sets the current function being built.

Parameters

function: The ValueTag of the new function to be set as the current function

pub fn get_current_insert_block(&self) -> Option<BasicBlockTag>

Retrieves the current block to insert after

Returns

Returns the tag of the basic block which is set as the block to insert after currently. Could be None if this is called when no basic block is currently set.

pub fn set_current_insert_block(&mut self, to_set: BasicBlockTag)

Sets the current block to insert after

Parameters

to_set: The BasicBlockTag of the new baic block to be set as the insert block

pub fn get_break_continue_target(&self) -> Option<Vec<BasicBlockTag>>

Retrieves the current break and potentially the current continue target

Returns

Returns a vector of basic block tags where index 0 is the break target and index 1 is the continue target (if one exists, otherwise the vector will only have index 0). This may return None if an error occurred getting the targets from the target stack.

pub fn push_break_continue_target( &mut self, break_block_tag: BasicBlockTag, continue_block_tag: BasicBlockTag, )

Pushes a tag for break; and a tag for continue; on the target stack.

Parameters

break_block_tag: The BasicBlockTag of the tag to branch to when break; is encountered.

continue_block_tag: The BasicBlockTag of the tag to branch to when continue; is encountered.

pub fn push_break_target(&mut self, break_block_tag: BasicBlockTag)

Pushes only a tag for break; on the target stack.

Parameters

break_block_tag: The BasicBlockTag of the tag to branch to when break; is encountered.

pub fn pop_target(&mut self)

Pops current break continue target off the stack

pub fn get_next_label_id(&mut self) -> String

Gets and increments the next ID number as a string for differentiating labels

Returns

Returns a string representing a unique ID number, advancing the current number that produced it by 1 so the next time this function is called it also returns a unique ID number.

Examples

/* IR is generated in different sections using labels, and labels must be unique so the code goes to
the right places. This allows you to use a system like */
//let label: String = "some_label_nameID".to_string();
//let next_id: String = self.get_next_label_id();
//let new_label_name = label + &next_id;
/* to create a new unique label name stored in new_label_name. For example if the next ID 
was 42 then this label would be some_label_nameID42, and if a new label needed to be generated
in this part of IR then it would do the same thing and be named some_label_nameID43, ensuring
all label names are unique. */

pub fn make_new_store_table(&mut self)

Makes a new current table in the store

Examples

/* When we enter a new scope, we need to create a new table in the store
to correctly generate IR such that variables only exist in their
current scope and inner scopes. When we do get into a new scope, use*/
//self.make_new_store_table();
/* to create a new table for this scope. */

pub fn delete_store_table(&mut self)

Deletes the current table in the store

Examples

/* When we leave a scope, we need to delete the current tab;e in the store
to correctly generate IR such that variables only exist in their
current scope and inner scopes, and when we return to an outer scope the 
variables in the current scope no longer exist. When we do get out of a
current scope, use*/
//self.delete_store_table();
/* to delete the current table for this scope. */

pub fn search_store_table(&self, name: String) -> ValueTag

Searches the store table for a variable with a name

Parameters

name: A String name of a variable to get an allocation tag for.

Returns

Returns the allocation tag of the variable associated with the given string variable name.

Examples

/* To retrieve a variable's value first you must get its allocation tag, use */
//let var_name: String = /* some variable name that exists in the current scope or an outer scope */
//let alloca_tag = self.search_store_table(var_name);
/* to do this. */

pub fn add_tag_to_store_table(&mut self, name: String, tag: ValueTag)

adds an allocation tag to the current store table

Parameters

name: A String name of a variable to set an allocation tag for.

tag: A ValueTag allocation tag of the variable to store in the table

Examples

/* When we create a new variable, it creates an allocation tag pointing that variable
that we must keep track of if we want to use that variable or reassign it in the future.
To do this, we associate variable names with tags in the store so we can retrieve
these allocation tags later. Use */
//let var_name: String = /* some variable name that we want to create*/
//let alloca_tag = /* allocation tag of the variable we've created with IR generation */
//self.add_tag_to_store_table(var_name, alloca_tag);
/* to do this. */

pub fn get_current_block(&mut self) -> Option<BasicBlockTag>

Retrieves the current basic block the builder is pointing into.

Returns

Returns the BasicBlockTag of the current basic block that the builder is working in. May be None if called before builder is positioned.

Examples

/* Inside the IR generation loop it may be necessary to get the tag of the current block for use in
other functions. Use*/
//let current_bb_tag = self.get_current_block();
/* to do this. */

pub fn increment_stack_pointer(&mut self)

Increments the stack pointer.

Examples

/* It will be necessary to increment the stack pointer when we need to move
 to a new scope. Use*/
//self.increment_stack_pointer();
/* to do this. */

pub fn decrement_stack_pointer(&mut self)

Decrements the stack pointer.

Examples

/* It will be necessary to decrement the stack pointer when we need to move back
 to a previous scope. Use*/
//self.decrement_stack_pointer();
/* to do this. */

pub fn reset_stack_pointer(&mut self)

Resets the stack pointer to zero.

Examples

/* Sometimes it could be necessary to reset the stack pointer to 0 if we need to move
 to a new STS. Use*/
//self.reset_stack_pointer();
/* to do this. */

pub fn get_stack(&self) -> Option<SymbolTableStack>

Retrieves a clone of the current symbol table stack.

Returns

Returns a clone of the current symbol table stack (may be None if called before the STS is initialized).

Examples

/* Inside the IR generation loop it may sometimes be necessary to retrieve
the current STS. Use*/
//let sts = self.get_stack();
/* to do this. */

pub fn set_stack(&mut self, new_sts: SymbolTableStack)

Sets the current symbol table stack to a new STS.

Parameters

new_sts: The new SymbolTableStack to be set as the STS for this IRGenerator.

Examples

/* with new_stack as some STS that is desired to be the current STS */
//self.set_stack(new_stack);

pub fn get_stack_pointer(&self) -> usize

Retrieves the current position of the stack pointer within the symbol table stack.

Returns

Returns a usize representing the current stack pointer.

Examples

/* when we need to access the STS, the way to find which stack to access is by 
using */
//let stack_pointer = self.get_stack_pointer();
/* to find the current value of the stack pointer, and access the STS with stack_pointer. */

pub fn generate_ir(&mut self, input: Module) -> ModuleTag

Generates LLVM IR from a given module by processing its AST. This method iterates over the module’s elements, generating IR for each and managing the state transitions required.

Parameters

input: A Module Containing an AST and an STS.

Returns

Returns a ModuleTag containing the constructed module when IR generation is complete.

pub fn ir_router(&mut self, node: &ASTNode) -> Result<Option<Tag>, ErrorType>

Routes the generation of LLVM IR based on the type of AST node encountered.

Parameters

node: A reference to an ASTNode to generate IR for.

Returns

Returns a Result<Option<Tag>, ErrorType> containing the tag of the data created by IR generation (may be None) or an error if there was a problem during IR generation.

Errors

Returns an error if there was a problem during IR generation such as a malformed node.

impl IRGenerator

pub fn generate_fn_declaration_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a function declaration.

Parameters

node: A reference to an ASTNode to generate IR for a function declaration.

Returns

Returns a Result<Option<Tag>, ErrorType> containing the Tag of this function if generation went smoothly or an Error if there was a problem generating the function.

Errors

Returns an ErrorType if generation failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a function declaration from */
//let result = self.generate_fn_declaration_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain the Tag that houses the 
function's ValueTag. */

pub fn generate_block_exp( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a block expression.

Parameters

node: A reference to an ASTNode to generate IR for a block expression.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the block expression.

Errors

Returns an ErrorType if generation failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a block expression from */
//let result = self.generate_block_exp(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_do_while_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a do while loop.

Parameters

node: A reference to an ASTNode to generate IR for a do while loop.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the do while.

Errors

Returns an ErrorType if generation failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a do while loop from */
//let result = self.generate_do_while_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_while_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a while loop.

Parameters

node: A reference to an ASTNode to generate IR for a while loop.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the while.

Errors

Returns an ErrorType if generation failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a while loop from */
//let result = self.generate_while_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_for_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a for loop.

Parameters

node: A reference to an ASTNode to generate IR for a for loop.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the for loop.

Errors

Returns an ErrorType if generation failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a for loop from */
//let result = self.generate_for_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_if_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for an if statement.

Parameters

node: A reference to an ASTNode to generate IR for an if statement.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the if statement.

Errors

Returns an ErrorType if generation failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate an if statement from */
//let result = self.generate_for_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

impl IRGenerator

pub fn generate_assignment_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for an assignment

Parameters

node: A reference to an ASTNode to generate IR for an assignment.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the initialization.

Errors

Returns an ErrorType if generation of this assignment failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate an assignment from */
//let result = self.generate_assignment_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_initialization_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for an initialization

Parameters

node: A reference to an ASTNode to generate IR for an initialization.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the initialization.

Errors

Returns an ErrorType if generation of this initialization failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate an initialization from */
//let result = self.generate_initialization_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_break_ir( &mut self, _node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a break statement.

Parameters

node: A reference to an ASTNode to generate IR for a break statement.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the break.

Errors

Returns an ErrorType if generation of this break failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a break statement from */
//let result = self.generate_break_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_continue_ir( &mut self, _node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a continue statement.

Parameters

node: A reference to an ASTNode to generate IR for a continue statement.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the continue.

Errors

Returns an ErrorType if generation of this continue failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a continue statement from */
//let result = self.generate_continue_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_return_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a return statement.

Parameters

node: A reference to an ASTNode to generate IR for a return statement.

Returns

Returns a Result<Option<Tag>, ErrorType> containing None if generation went smoothly or an Error if there was a problem generating the return.

Errors

Returns an ErrorType if generation of this return failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a return statement from */
//let result = self.generate_return_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain None. */

pub fn generate_variable_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for recalling a variable

Parameters

node: A reference to an ASTNode to generate IR for a variable recall statement.

Returns

Returns a Result<Option<Tag>, ErrorType> containing the Tag of the recalled variable container if generation went smoothly or an Error if there was a problem in generation.

Errors

Returns an ErrorType if generation failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a recall statement from */
//let result = self.generate_variable_ir(&a_node);
/* check if type_result was Ok or Err, if Ok, it will contain a Tag that houses
the allocation tag of the recalled variable container. */

impl IRGenerator

pub fn generate_data_type_ir( &mut self, data_type: &DataType, ) -> Result<Option<Tag>, ErrorType>

Generates an LLVM type tag for a data type.

Parameters

data_type: A reference to a DataType representing the type that should be created in the resource pools.

Returns

Returns a Result<Option<Tag>, ErrorType> containing the constructed tag of the generated type if successful, or an ErrorType if there was an error generating this type tag.

Errors

Returns an ErrorType if generation of this type tag failed.

Examples

//let d_type: DataType = /* Some DataType we want a data type tag for */
//let type_result = self.generate_data_type_ir(&d_type);
/* check if type_result was Ok or Err, if Ok extract the TypeTag from
the Tag and use this for other functions. */

pub fn generate_literal_ir( &mut self, node: &ASTNode, ) -> Result<Option<Tag>, ErrorType>

Generates LLVM IR for a literal.

Parameters

node: A reference to an ASTNode to generate IR for a literal value from.

Returns

Returns a Result<Option<Tag>, ErrorType> containing the tag of the generated literal value or an Error if there was a problem generating the literal.

Errors

Returns an ErrorType if generation of this literal failed.

Examples

//let a_node: ASTNode = /* Some ASTNode we want to generate a literal value from */
//let result = self.generate_literal_ir(&a_node);
/* check if type_result was Ok or Err, if Ok extract the ValueTag from
the Tag and use this for other functions. */