Enum ir::mir::MIRInstructions

#[repr(u8)]pub enum MIRInstructions {
    Module,
    EndModule,
    Fun,
    EndFun,
    FunParam,
    Integer,
    Float,
    String,
    Bool,
    Unit,
    ObjInit,
    Drop,
    Ref,
    Move,
    Copy,
    HeapAlloc,
    StackAlloc,
    Lateinit,
    ObjMut,
    Halt,
}

Variants

Module

Module start

EndModule

End module

Fun

Function start. At the beginning is where local variable preallocation will occur.

EndFun

End function. This is where all drops to local variables and any references or values passed as arguments will occur.

FunParam

Integer

Integer literal

Float

Float literal

String

String literal

Bool

Boolean literal

Unit

Unit type

ObjInit

Initialize object name with mutability. An allocation instruction must precede this with the size of the object. Following this will be a call to the initializer and its mutability.

Drop

Drop name. This can either be a value or a reference. The drop mechanism is smart. If what is being dropped is a reference, the reference counter will decrement the count for object name.

Ref

Create reference for refee. High level references to properties will result in this instruction.

Move

Move name. A single reference to a local variables will result in this instruction.

Copy

Copy name. Where n is the number of references to a local variable, all references until n-1 will result in this instruction, whereas the final reference to a local variable will result in a Move instruction.

HeapAlloc

Heap allocation of size for object name

StackAlloc

Stack allocation of size for object name. Either an object contruction or a lateinit instruction must proceed this.

Lateinit

Uninitialized/late initializer. This is used for leaving an resource empty until further notice. For immutable objects, this grants one free initial mutation for initialization, to which all subsequent mutations will become invalid. None is the placehold value, so instead of an unsafe empty place in memory, None will fill the emptyness. None is an object that can be stretched to fit any place whatsoever, and will simply just be garbage data. The syntax for this is: let something: A = None

ObjMut

Mutate object name. An expression must proceed this instruction.

Halt

Halt compiler